+++ app/src/main/kotlin/com/simplemobiletools/calculator/operation/base/Operation.kt
@@ -1,7 +1,5 @@
 package com.simplemobiletools.calculator.operation.base
 
-import java.math.BigDecimal
-
 interface Operation {
-    fun getResult(): BigDecimal
+    fun getResult(): Double
 }
